allows user to enter floating-point values for lineWidth

+++ src/main/java/com/bibounde/vprotovisdemo/areachart/AreaChartPage.java
@@ -175,7 +175,7 @@ public class AreaChartPage implements Page {
             chart.setMarginBottom(10d);
         }
         
-        Integer lineWidth = this.dimensionPanel.getLineWidth();
+        Double lineWidth = this.dimensionPanel.getLineWidth();
         if (lineWidth != null) {
             chart.setLineWidth(lineWidth);
             this.sourceCodeMap.put("lineWidth", lineWidth);
