+++ src/main/java/com/bibounde/vprotovisdemo/piechart/PieChartPage.java
@@ -181,7 +181,7 @@ public class PieChartPage implements Page {
             chart.setHighlightOffset(10d);
         }
         
-        Integer lineWidth = this.dimensionPanel.getLineWidth();
+        Double lineWidth = this.dimensionPanel.getLineWidth();
         if (lineWidth != null) {
             chart.setLineWidth(lineWidth);
             this.sourceCodeMap.put("lineWidth", lineWidth);
